GUI investment strategies

+++ src/models/FixedAmountStrategy.java
@@ -7,12 +7,12 @@ import java.util.Map;
 public class FixedAmountStrategy<T> implements PortfolioOperation<T> {
 
   private final String portfolioName;
-  private final int investmentAmount;
+  private final float investmentAmount;
   private final LocalDate date;
   private final HashMap<String, Float> stocks;
   private final float commission;
 
-  public FixedAmountStrategy(String portfolioName, int investmentAmount,
+  public FixedAmountStrategy(String portfolioName, float investmentAmount,
                              LocalDate date, HashMap<String, Float> stocks, float commission) {
     this.portfolioName = portfolioName;
     this.investmentAmount = investmentAmount;
